Read changeset IDs from the repo pushlog
We receive a single push event from the hgpush service for each push to
the repo.  However, pushes could contain dozens of new commits behind
merges in the newly pushed DAG.

Mozilla hg repos record a list of newly pushed changesets in a custom
pushlog extension.  This means the work of finding new changesets has
been done for us.  For each hgpush event, read the pushlog for that
event and process all of the changesets found within.